Added new ammotypes, HEAT and APHE. HEAT uses it's explosive filler to send a molten jet of metal at very high velocities, APHE works as an AP shell until it fails to penetrate, then the HE filler explodes.
Added Engines and Gearbox system, EXPERIMENTAL. You link engines to the gearboxes, then the gearboxes to your wheels, with the same right click as the ammo links. The engine will provide power to the wheels with an appropriate power and torque curve.

Rewrote HE explosion effect to separate the shockwave from the main explosion, and be aware of the mattype it hits

git-svn-id: http://acf.googlecode.com/svn/trunk@46 d1139d92-4d4e-2ec3-fc3e-6e4027237e5b