Beta test. Cannon penetration has been buffed ~25%. HE power has been increased by 30%.
50mmC 92mm > 107mm
75mmC 138mm > 150mm
100mmC 175mm > 200mm
120mmC 230mm > 250mm
140mmC 280mm > 300mm

This is a test to see how more penetration does. Obviously it will outdate your current vehicles, but your vehicles will still have the same ammunition as before.
This test is to discourage slugfests further and force people to use some tactics. ACF build tech has advanced in the past three years and the old penetration values just aren't cutting it anymore.
Other solutions to fix this are a bit more intrusive.