--Health divisor changed from 300 > 225 (25% increase in overall prop health) --General ammo capacity nerfed 40% --HE power nerfed 20% (10000 > 8000, default was 6000)
These changes are to try to make roadwheels stronger (less HE power, more health) and force tanks to go larger to carry more ammo or use smaller engines to fit ammo in existing vehicles. There is currently no incentive to use a smaller engine in a tank.

Rough change, but gmod is about building new things all the time. Remember, this is a beta svn for a reason.