Added high grade fuels
Fuel is not required for most engines.  When used, fuel will boost
engine torque (and consequently power) by 25%.  Making a tank active,
and putting it on refuel duty will cause the tank to supply fuel to
other nearby tanks.  Fuel consumption is based on engine type and engine
power.  RPM and throttle also affect fuel consumption.  Diesels and
turbines are efficient, while wankels and radials are not.